Refactor articles and comments modules with schema and DTO separation
