+++ src/articles/dto/index.ts
@@ -1,0 +1,5 @@
+export * from './create-article.dto';
+export * from './update-article.dto';
+export * from './article-response.dto';
+export * from './list-articles-query.dto';
+export * from './list-articles-response.dto';
