+++ src/articles/dto/list-articles-query.dto.ts
@@ -1,0 +1,12 @@
+import { Type, type Static } from '@sinclair/typebox';
+import { ArticleFeedQueryDto } from './article-feed-query.dto';
+
+export const ListArticlesQueryDto = Type.Composite([
+  ArticleFeedQueryDto,
+  Type.Object({
+    tag: Type.Optional(Type.String()),
+    author: Type.Optional(Type.String()),
+    favorited: Type.Optional(Type.String()),
+  }),
+]);
+export type ListArticlesQueryDto = Static<typeof ListArticlesQueryDto>;
