+++ src/articles/dto/list-articles-response.dto.ts
@@ -1,0 +1,8 @@
+import { Type, type Static } from '@sinclair/typebox';
+import { ArticleResponseDto } from './article-response.dto';
+
+export const ListArticlesResponseDto = Type.Object({
+  articles: Type.Array(ArticleResponseDto),
+  articlesCount: Type.Number(),
+});
+export type ListArticlesResponseDto = Static<typeof ListArticlesResponseDto>;
