+++ src/articles/schema/index.ts
@@ -1,0 +1,4 @@
+export * from './articles.schema';
+export * from './article-relations.schema';
+export * from './comment-relations.schema';
+export * from './favorite-articles.schema';
