+++ src/comments/comments.repository.ts
@@ -1,5 +1,5 @@
 import type { Database } from '@/database.providers';
-import { articles, comments } from '@articles/articles.model';
+import { articles, comments } from '@/articles/schema/favorite-articles.schema';
 import { and, desc, eq } from 'drizzle-orm';
 import type { CommentToCreate } from './comments.schema';
