+++ src/comments/dto/create-comment.dto.ts
@@ -1,0 +1,8 @@
+import { Type, type Static } from '@sinclair/typebox';
+
+export const CreateCommentDto = Type.Object({
+  comment: Type.Object({
+    body: Type.String({ minLength: 1 }),
+  }),
+});
+export type CreateCommentDto = Static<typeof CreateCommentDto>;
