+++ src/comments/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './create-comment.dto';
+export * from './comment-response.dto';
