+++ src/comments/schema/index.ts
@@ -1,0 +1,2 @@
+export * from './comments.schema';
+export * from './comment-relations.schema';
