+++ src/users/users.model.ts
@@ -1,4 +1,7 @@
-import { articles, favoriteArticles } from '@/articles/articles.model';
+import {
+  articles,
+  favoriteArticles,
+} from '@/articles/schema/favorite-articles.schema';
 import { relations, sql } from 'drizzle-orm';
 import {
   date,
